requirements.txt
langchain==0.1.13
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.33
openai==1.25.2
faiss-cpu==1.7.4
altair==4
tiktoken==0.5.2
langchain-openai == 0.1.6
transformers == 4.40.1
youtube-transcript-api == 0.6.2
chromadb == 0.3.29
pydantic == 1.10.15
beautifulsoup4 == 4.12.2
streamlit-google-auth == 1.1.8

